b1.0.1 hotfix

+++ laravel/app/Http/Controllers/PlaceController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Place;
 use App\Models\File;
+use App\Models\Favorite;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 
@@ -16,8 +17,10 @@ class PlaceController extends Controller
      */
     public function index()
     {
+        $places = Place::withCount('favorites')->get();
+
         return view("places.index", [
-            "places" => Place::all()
+            "places" => $places
         ]);
     }
 
@@ -46,6 +49,7 @@ class PlaceController extends Controller
             'upload'      => 'required|mimes:gif,jpeg,jpg,png,mp4|max:2048',
             'latitude'    => 'required',
             'longitude'   => 'required',
+            'visibility'  => 'required|exists:visibilities,id',
         ]);
         
         // Obtenir dades del formulari
@@ -54,6 +58,7 @@ class PlaceController extends Controller
         $upload      = $request->file('upload');
         $latitude    = $request->get('latitude');
         $longitude   = $request->get('longitude');
+        $visibility  = $request->get('visibility');
 
         // Desar fitxer al disc i inserir dades a BD
         $file = new File();
@@ -63,17 +68,20 @@ class PlaceController extends Controller
             // Desar dades a BD
             Log::debug("Saving place at DB...");
             $place = Place::create([
-                'name'        => $name,
-                'description' => $description,
-                'file_id'     => $file->id,
-                'latitude'    => $latitude,
-                'longitude'   => $longitude,
-                'author_id'   => auth()->user()->id,
+                'name'          => $name,
+                'description'   => $description,
+                'file_id'       => $file->id,
+                'latitude'      => $latitude,
+                'longitude'     => $longitude,
+                'visibility_id' => $visibility,
+                'author_id'     => auth()->user()->id,
             ]);
             \Log::debug("DB storage OK");
             // Patró PRG amb missatge d'èxit
             return redirect()->route('places.show', $place)
-                ->with('success', __('Place successfully saved'));
+                ->with('success', __(':resource successfully saved', [
+                    'resource' => __('resources.place')
+                ]));
         } else {
             \Log::debug("Disk storage FAILS");
             // Patró PRG amb missatge d'error
@@ -90,10 +98,14 @@ class PlaceController extends Controller
      */
     public function show(Place $place)
     {
+        // Favorites counter
+        $place->loadCount('favorites');
+        
         return view("places.show", [
             'place'  => $place,
             'file'   => $place->file,
             'author' => $place->user,
+            'numFavorites' => $place->favorites_count,
         ]);
     }
 
@@ -105,6 +117,9 @@ class PlaceController extends Controller
      */
     public function edit(Place $place)
     {
+        // Comprovar si l'usuari/a pot editar (PlacePolicy)
+        $this->authorize('update', $place);
+
         return view("places.edit", [
             'place'  => $place,
             'file'   => $place->file,
@@ -121,6 +136,9 @@ class PlaceController extends Controller
      */
     public function update(Request $request, Place $place)
     {
+        // Comprovar si l'usuari/a pot editar (PlacePolicy)
+        $this->authorize('update', $place);
+
         // Validar dades del formulari
         $validatedData = $request->validate([
             'name'        => 'required',
@@ -128,6 +146,7 @@ class PlaceController extends Controller
             'upload'      => 'nullable|mimes:gif,jpeg,jpg,png,mp4|max:2048',
             'latitude'    => 'required',
             'longitude'   => 'required',
+            'visibility'  => 'required|exists:visibilities,id',
         ]);
         
         // Obtenir dades del formulari
@@ -136,20 +155,24 @@ class PlaceController extends Controller
         $upload      = $request->file('upload');
         $latitude    = $request->get('latitude');
         $longitude   = $request->get('longitude');
+        $visibility  = $request->get('visibility');
 
         // Desar fitxer (opcional)
         if (is_null($upload) || $place->file->diskSave($upload)) {
             // Actualitzar dades a BD
             Log::debug("Updating DB...");
-            $place->name        = $name;
-            $place->description = $description;
-            $place->latitude    = $latitude;
-            $place->longitude   = $longitude;
+            $place->name          = $name;
+            $place->description   = $description;
+            $place->latitude      = $latitude;
+            $place->longitude     = $longitude;
+            $place->visibility_id = $visibility;
             $place->save();
             \Log::debug("DB storage OK");
             // Patró PRG amb missatge d'èxit
             return redirect()->route('places.show', $place)
-                ->with('success', __('Place successfully saved'));
+                ->with('success', __(':resource successfully updated', [
+                    'resource' => __('resources.place')
+                ]));
         } else {
             // Patró PRG amb missatge d'error
             return redirect()->route("places.create")
@@ -165,12 +188,59 @@ class PlaceController extends Controller
      */
     public function destroy(Place $place)
     {
+        // Comprovar si l'usuari/a pot eliminar (PlacePolicy)
+        $this->authorize('delete', $place);
+        
         // Eliminar place de BD
         $place->delete();
+
         // Eliminar fitxer associat del disc i BD
         $place->file->diskDelete();
+
         // Patró PRG amb missatge d'èxit
         return redirect()->route("places.index")
-            ->with('success', 'Place successfully deleted');
+            ->with('success', __(':resource successfully deleted', [
+                'resource' => __('resources.place')
+            ]));
+    }
+
+    /**
+     * Add favorite place
+     *
+     * @param  \App\Models\Place  $place
+     * @return \Illuminate\Http\Response
+     */
+    public function favorite(Place $place) 
+    {
+        $fav = Favorite::create([
+            'user_id'  => auth()->user()->id,
+            'place_id' => $place->id
+        ]);
+
+        return redirect()->back()
+            ->with('success', __(':resource successfully saved', [
+                'resource' => __('resources.favorite')
+            ]));
+    }
+
+    /**
+     * Undo favorite
+     *
+     * @param  \App\Models\Place  $place
+     * @return \Illuminate\Http\Response
+     */
+    public function unfavorite(Place $place) 
+    {
+        $fav = Favorite::where([
+            ['user_id',  '=', auth()->user()->id],
+            ['place_id', '=', $place->id],
+        ])->first();
+        
+        $fav->delete();
+        
+        return redirect()->back()
+            ->with('success', __(':resource successfully deleted', [
+                'resource' => __('resources.favorite')
+            ]));
     }
 }
